Update 5.15.9
The code is a bit jumbled right now, but I release this mostly as a
major code improvement. Loading time is down among other things.

There are a few bugs I know of (or I think so anyway), and an
unimplemented feature, which I'm pretty sure I'll have done by the next
release.