Update 5.90.8850
This update is mostly bug and stability fixes, and normally I'd put it
off until I released the final version but a LOT of them are critical
bugs so I figured I should release.

I will not be focusing on code optimization or features until update
6.0.

The next update should be the official 6.0 release.