6.50 (Rev: 38499)
So I've been gone quite a while and figured to post this only to let
anyone who cares know very little in the way of updates or anything
really will be released on this project in any regular manner. E.G. I'll
get to it when I get to it.

Otherwise just minor architecture fixes. I'm sure other stuff too but
it's been so long I forget.