6.60 (Rev: 32398)
A few significant improvements have been made to secondary parts of the
program such as first time use and documentation. I am working on both
help documentation and a sync feature at the moment so things are pretty
messy, but I figured I'd post just so I'm not too much of a ghost (as I
have been lately).