6.65 (Rev: hotfix)
So in the update I made ~1 week ago I think I did something to make the
controls unbind from their respective functions. I believe I've brought
it up to par. This update is on the off-chance there is someone out
there beside me using this on a regular basis and got the last update
and couldn't do anything.

I do not recommend this update otherwise. I am currently working on
memory-leaking issues as well as the code is rather messy due to
preoccupation with other facets of the program (including the new sync
feature that -still- isn't finished.