6.75 [Rev: 13176]
So, normally I'd wait until I had a significant portion of what I
started be finished. In this case, I have finished some, but left a lot
to be finished (hopefully by update 7). The major features that need to
be finished are sync and the help support.

I'm thinking I'll get to help right before releasing update 7, as for
sync I am unsure if I'll be able to finish by then. Sync has been very
troublesome....

Anyway, this is more-or-less stable.